Updates to the front end search feature
