+++ code/react-app/src/App.js
@@ -10,23 +10,19 @@ import Resources from "./pages/Resources";
 import NotFound from "./pages/NotFound";
 import Login from "./login/Login";
 import { useEffect, useState } from "react";
-import { Users } from "./pages/users";
-import "./pages/app.css";
-import Table from "./pages/Table";
+
 
 
 
    function App() {
-    const [query, setQuery] = useState("");
- const search = (data) =>{
-   return data.filter((item)=> item.name.toLowerCase().includes(query));
- };
+   
+ 
   
     return (
       <div>
         <Router>
           <Navbar />
-          <Login />
+          
           <Routes>
             <Route path="/" element={<Home />} />
             <Route path="/calculator" element={<Calculator />} />
@@ -37,15 +33,7 @@ import Table from "./pages/Table";
           </Routes>
         </Router>
 
-        <div className="app">
-       <input
-       type="text"
-         className="search"
-         placeholder="Search..."
-         onChange={(e) => setQuery(e.target.value)}
-       />
-     <Table data={search(Users)} />
-   </div>
+      
       </div>
     );
